오류 수정

+++ dungreed/views.py
@@ -8,7 +8,7 @@ def dps_test(request):
     for i in check_list:
         if not request.GET.get(i):
             return render(request, 'dungreed/dps.html')
-    datas=list(map(int,request.GET.dict().values()))
+    datas=list(map(float,request.GET.dict().values()))
     q = Calc().basic(datas)
     print(datas)
     return render(request, 'dungreed/dps.html', {'q':q})
@@ -16,4 +16,3 @@ def dps_test(request):
 # https://tutorial.djangogirls.org/ko/extend_your_application/
 # 에서 404 뭐시기 db에서 데이터 못찾으면 할거 그거 있음
 # view에서는 처리를 해야함 (MVC의 Control)
-#
